fix saving the right margin
svn path=/trunk/koffice/; revision=941694

+++ kexi/plugins/reportspgz/backend/wrtembed/reportdesigner.cpp
@@ -650,7 +650,7 @@ QDomDocument ReportDesigner::document()
     margin.appendChild(doc.createTextNode(QString::number(_bottomMargin->value().toDouble())));
     root.appendChild(margin);
     margin = doc.createElement("rightmargin");
-    margin.appendChild(doc.createTextNode(QString::number(_rightMargin->value().toInt())));
+    margin.appendChild(doc.createTextNode(QString::number(_rightMargin->value().toDouble())));
     root.appendChild(margin);
     margin = doc.createElement("leftmargin");
     margin.appendChild(doc.createTextNode(QString::number(_leftMargin->value().toDouble())));
